moved components out of atomic structure as most are for the landing page
