+++ src/pages/features/ComparisonTable.tsx
